chore: fixed unit tests

+++ lambda/app/src/routes.ts
@@ -18,6 +18,7 @@ import {
 } from './controllers/team';
 import {
   findOrCreateUser,
+  isAllowedToManageRoles,
   listClientRolesByUsers,
   listUsersByRole,
   updateProfile,
@@ -33,6 +34,7 @@ import {
   deleteRequest,
   updateRequestMetadata,
   getIntegrations,
+  isAllowedToDeleteIntegration,
 } from './controllers/requests';
 import { getInstallation, changeSecret } from './controllers/installation';
 import { searchKeycloakUsers } from './controllers/keycloak';
@@ -216,6 +218,12 @@ export const setRoutes = (app: any) => {
   app.delete(`${BASE_PATH}/requests`, async (req, res) => {
     try {
       const { id } = req.query || {};
+
+      const authorized = await isAllowedToDeleteIntegration(req.session as Session, id);
+
+      if (!authorized)
+        return res.status(401).json({ success: false, message: 'You are not authorized to delete this integration' });
+
       const result = await deleteRequest(req.session as Session, req.user, Number(id));
       res.status(200).json(result);
     } catch (err) {
@@ -351,6 +359,11 @@ export const setRoutes = (app: any) => {
 
   app.post(`${BASE_PATH}/keycloak/bulk-roles`, async (req, res) => {
     try {
+      const authorized = await isAllowedToManageRoles(req.session as Session, req.body.integrationId);
+
+      if (!authorized)
+        return res.status(401).json({ success: false, message: 'You are not authorized to create role' });
+
       const result = await bulkCreateRole((req.session as Session).user.id, req.body);
       res.status(200).json(result);
     } catch (err) {
@@ -360,6 +373,11 @@ export const setRoutes = (app: any) => {
 
   app.post(`${BASE_PATH}/keycloak/delete-role`, async (req, res) => {
     try {
+      const authorized = await isAllowedToManageRoles(req.session as Session, req.body.integrationId);
+
+      if (!authorized)
+        return res.status(401).json({ success: false, message: 'You are not authorized to delete role' });
+
       const result = await deleteRoles((req.session as Session).user.id, req.body);
       res.status(200).json(result);
     } catch (err) {
@@ -434,8 +452,10 @@ export const setRoutes = (app: any) => {
     try {
       const { id } = req.params;
       const authorized = await userIsTeamAdmin(req.user, id);
+
       if (!authorized)
         return res.status(401).json({ success: false, message: 'You are not authorized to edit this team' });
+
       const result = await updateTeam(req.user, id, req.body);
       res.status(200).json(result);
     } catch (err) {
@@ -447,8 +467,10 @@ export const setRoutes = (app: any) => {
     try {
       const { id } = req.params;
       const authorized = await userIsTeamAdmin(req.user, id);
+
       if (!authorized)
         return res.status(401).json({ success: false, message: 'You are not authorized to edit this team' });
+
       const result = await addUsersToTeam(id, req.user.id, req.body);
       res.status(200).json(result);
     } catch (err) {
